[Shunting Yard] Use Generics.Collections
There is limited documentation on the Generics.Collectionis library, but it is apparently part of the runtime library as of FPC 3.2.2 or so. This seems more type safe than the previous method I was using that converted everything to TObject. Another possible option is the fgl that has more documention, but Generics.Collections seemslike it will be easier with which to work.