Replace the help.text file with a separate unit
The goal is for this to be one self-contained executable file. The text file would have to be sent along to be usable, failing the goal. There may be an option to create a resource file with the information and have it bundled into the executable. That approach would have taken a lot more time an energy to figure out how to implement it. Creating a separate unit for printing the help information keeps it out of the way and still easy to update if needed.