Use custom StringQueue
Replace the Generics.Collection TQueue with a custom StringQueue. When compiled in debug mode with TQueue from Generics.Collection, an error would be raised when trying to iterate through the queue. It claimed there was an illegal type cast. A work around was used that utilized an iterator, but it was not in the spirit of having a queue.