Update ShuntingYardParser
Update ShuntingYardParser to be more consistent in how objects are freed.
In this case, the output queue is only produced by ShuntingYardParser, it is not
owned by the class and it is not the final user. It makes sense for the consumer
of the output queue to own it and be responsible for freeing it.